feat: Complete payment API deployment with AKS and SQL Server
- Implemented payment processing API with charge, capture, and cancel operations
- Added idempotency support to prevent duplicate transactions
- Deployed SQL Server 2022 in-cluster with persistent storage (8GB)
- Configured Kubernetes deployment with auto-scaling (2-10 pods)
- Set up LoadBalancer service with external IP (4.213.208.156)
- Added comprehensive documentation (README, API endpoints, deployment status)
- Implemented Docker containerization with multi-stage builds
- Fixed Entity Framework data types (Guid IDs, long amounts in cents)
- Added health monitoring and readiness probes
- Protected sensitive files with updated .gitignore
- Cleaned up temporary resources and redundant documentation

+++ WebApiProject/Models/ChargeRequest.cs
@@ -2,7 +2,7 @@ namespace WebApiProject.Models;
 
 public class ChargeRequest
 {
-  public decimal Amount { get; set; }
+  public long Amount { get; set; } // Amount in cents (e.g., 1000 = $10.00)
   public string Currency { get; set; } = "USD";
   public string? Description { get; set; }
   public string? CustomerId { get; set; }
